requirements.txt
Flask==2.3.3
yagmail==0.15.293
pandas==2.1.0
openpyxl==3.1.2
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3 